fix: move create_redirect_response to auth_provider
